Refactor re-exports
