+++ packages/core/src/index.ts
@@ -20,10 +20,11 @@ export {
 } from "./interpreter/index.js";
 export { createStore } from "./store/index.js";
 export { createRenderer } from "./renderer/index.js";
-export { createReduce, toPublicState } from "./model/index.js";
 export {
+  createReduce,
+  toPublicState,
   createCarouselReduce,
   toCarouselPublicState,
   type CarouselConfig,
   type CarouselPublicState,
-} from "./model/carousel.js";
+} from "./model/index.js";
